requirements.txt
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
geoip2==2.3.0
gitdb==0.6.4
GitPython==0.3.2rc1
httplib2==0.9.2
ipaddr==2.1.11
lxml==3.1b1
maxminddb==1.2.0
mygengo==1.3.3
netaddr==0.7.18
oauth2==1.9.0.post1
oauthlib==1.1.1
pycrypto==2.6.1
pydns==2.3.6
pyenchant==1.6.5
pygeoip==0.3.2
python-mpd2==0.5.5
pytz==2016.4
pywhois==0.1
requests==2.10.0
requests-oauthlib==0.6.1
simplejson==3.8.2
six==1.10.0
smmap==0.9.0
timezones==1.9.6
tweepy==3.5.0
